CODE/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# $ pip install -r <this_file>
matplotlib==3.7.0
numpy==1.23.5
pandas==1.2.1
progressbar2==4.2.0
python==3.8.2
scipy==1.10.0
